# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.2.0
  description: Seeds API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /evaluator/assignments:
    get:
      operationId: listMyAssignments
      tags: [evaluator]
      summary: List the current evaluator's assignments
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EvaluatorAssignmentList"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /evaluator/applications/{id}:
    get:
      operationId: getEvaluatorApplication
      tags: [evaluator]
      summary: Get an application the current evaluator is assigned to (with own evaluation only)
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EvaluatorApplicationDetail"
        "403":
          description: Not assigned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /evaluator/applications/{id}/evaluations:
    post:
      operationId: submitEvaluation
      tags: [evaluator]
      summary: Submit or update the current evaluator's evaluation for the given stage
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmitEvaluationBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Evaluation"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Not assigned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error: { type: string }
      required: [error]
...
    ApplicationStatus:
      type: string
      enum:
        - submitted
        - reviewing
        - interview
        - accepted
        - rejected
        - waitlisted
        - withdrawn
    ApplicationLifecycleStatus:
      type: string
      enum:
        - submitted
        - document_review
        - document_review_completed
        - interview
        - interview_scheduled
        - interview_completed
        - final_decision_made
        - withdrawn
    FinalDecision:
      type: string
      enum: [pending, accepted, rejected, waitlisted, withdrawn]
    EvaluationStage:
      type: string
      enum: [document_review, interview]
    AssignmentStatus:
      type: string
      enum: [assigned, in_progress, completed]
...
    Recommendation:
      type: string
      enum: [strong_accept, accept, hold, reject, strong_reject]
...
    Application:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        email: { type: string }
        phone: { type: string }
        school: { type: string }
        grade: { type: string }
        birthYear: { type: integer }
        interestArea: { type: string }
        motivation: { type: string }
        experience: { type: string }
        problemAwareness: { type: string }
        expectation: { type: string }
        privacyConsent: { type: boolean }
        status:
          $ref: "#/components/schemas/ApplicationStatus"
        applicationStatus:
          $ref: "#/components/schemas/ApplicationLifecycleStatus"
        finalDecision:
          $ref: "#/components/schemas/FinalDecision"
        adminNote:
          type: string
          nullable: true
        submittedAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
      required:
        - id
        - name
        - email
        - phone
        - school
        - grade
        - birthYear
        - interestArea
        - motivation
        - experience
        - problemAwareness
        - expectation
        - privacyConsent
        - status
        - applicationStatus
        - finalDecision
        - adminNote
        - submittedAt
        - updatedAt
...
    EvaluationAssignment:
      type: object
      properties:
        id: { type: integer }
        applicationId: { type: integer }
        evaluatorId: { type: integer }
        evaluatorName: { type: string }
        evaluatorEmail: { type: string }
        stage:
          $ref: "#/components/schemas/EvaluationStage"
        status:
          $ref: "#/components/schemas/AssignmentStatus"
        assignedAt: { type: string, format: date-time }
      required: [id, applicationId, evaluatorId, evaluatorName, evaluatorEmail, stage, status, assignedAt]
...
    Evaluation:
      type: object
      properties:
        id: { type: integer }
        applicationId: { type: integer }
        evaluatorId: { type: integer }
        evaluatorName: { type: string }
        stage:
          $ref: "#/components/schemas/EvaluationStage"
        motivationScore: { type: integer, nullable: true }
        problemAwarenessScore: { type: integer, nullable: true }
        initiativeScore: { type: integer, nullable: true }
        collaborationScore: { type: integer, nullable: true }
        fitScore: { type: integer, nullable: true }
        overallScore: { type: integer }
        recommendation:
          $ref: "#/components/schemas/Recommendation"
        comment:
          type: string
          nullable: true
        submittedAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
      required:
        - id
        - applicationId
        - evaluatorId
        - evaluatorName
        - stage
        - motivationScore
        - problemAwarenessScore
        - initiativeScore
        - collaborationScore
        - fitScore
        - overallScore
        - recommendation
        - comment
        - submittedAt
        - updatedAt
    SubmitEvaluationBody:
      type: object
      properties:
        stage:
          $ref: "#/components/schemas/EvaluationStage"
        motivationScore: { type: integer, minimum: 1, maximum: 5, nullable: true }
        problemAwarenessScore: { type: integer, minimum: 1, maximum: 5, nullable: true }
        initiativeScore: { type: integer, minimum: 1, maximum: 5, nullable: true }
        collaborationScore: { type: integer, minimum: 1, maximum: 5, nullable: true }
        fitScore: { type: integer, minimum: 1, maximum: 5, nullable: true }
        overallScore: { type: integer, minimum: 1, maximum: 5 }
        recommendation:
          $ref: "#/components/schemas/Recommendation"
        comment:
          type: string
          nullable: true
          maxLength: 5000
      required: [stage, overallScore, recommendation]
...
    EvaluatorAssignmentItem:
      type: object
      properties:
        assignmentId: { type: integer }
        applicationId: { type: integer }
        applicantName: { type: string }
        applicantSchool: { type: string }
        stage:
          $ref: "#/components/schemas/EvaluationStage"
        assignmentStatus:
          $ref: "#/components/schemas/AssignmentStatus"
        hasEvaluation: { type: boolean }
        assignedAt: { type: string, format: date-time }
      required: [assignmentId, applicationId, applicantName, applicantSchool, stage, assignmentStatus, hasEvaluation, assignedAt]
    EvaluatorAssignmentList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/EvaluatorAssignmentItem"
        total: { type: integer }
      required: [items, total]
    EvaluatorApplicationDetail:
      type: object
      properties:
        application:
          $ref: "#/components/schemas/Application"
        myAssignments:
          type: array
          items:
            $ref: "#/components/schemas/EvaluationAssignment"
        myEvaluations:
          type: array
          items:
            $ref: "#/components/schemas/Evaluation"
      required: [application, myAssignments, myEvaluations]