# lib/api-zod/src/generated/types/evaluatorAssignmentList.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Seeds API
 * OpenAPI spec version: 0.2.0
 */
import type { EvaluatorAssignmentItem } from "./evaluatorAssignmentItem";

export interface EvaluatorAssignmentList {
  items: EvaluatorAssignmentItem[];
  total: number;
}